api/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.2.0+cpu # Enable before creating the container image, comment below torch==2.2.0
fastapi
uvicorn
torch==2.2.0
pymilvus==2.3.6
sentence-transformers==2.3.1
flask-openapi3==3.0.2
langchainhub==0.1.14
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
cryptography==42.0.5
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
ibm-cos-sdk==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
ibm-db==3.2.1
ibm-watson-machine-learning==1.0.352
idna==3.6
importlib_metadata==7.0.2
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain-text-splitters==0.0.1
langsmith==0.1.27
lomond==0.3.3
marshmallow==3.21.1
matplotlib-inline==0.1.6
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
orjson==3.9.15
packaging==23.2
pandas==1.5.3
parso==0.8.3
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pypdfium2==4.28.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tornado==6.4
traitlets==5.14.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.1.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.18.1
